pyproject.toml
[project]
name = "5tran"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dbt-bigquery>=1.10.2",
    "dbt-core>=1.10.8",
    "fivetran>=0.7.0",
    "google-cloud-bigquery>=3.38.0",
    "google-generativeai>=0.8.5",
    "gradio>=5.49.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
